requirements.txt
Flask==2.3.3
python-telegram-bot==20.7
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.2
cryptography==41.0.7
urllib3==2.0.7
gunicorn==21.2.0